Fix test conflict.
