+++ service/src/test/java/bio/terra/workspace/app/controller/ControlledFlexibleResourceApiControllerTest.java
@@ -268,63 +268,4 @@ public class ControlledFlexibleResourceApiControllerTest extends BaseUnitTest {
     // Assert clone doesn't exist. There's no resource ID, so search on resource name.
     mockMvcUtils.assertNoResourceWithName(USER_REQUEST, workspaceId, destResourceName);
   }
-
-  @Test
-  void clone_copyResource() throws Exception {
-    UUID workspaceId = mockMvcUtils.createWorkspaceWithoutCloudContext(USER_REQUEST).getId();
-    UUID workspaceId2 = mockMvcUtils.createWorkspaceWithoutCloudContext(USER_REQUEST).getId();
-    var sourceFlex = createDefaultFlexResourceWithData(workspaceId, null);
-
-    String destResourcename = TestUtils.appendRandomNumber("dest-resource-name");
-    String destDescription = "new description";
-
-    ApiFlexibleResource clonedFlexResource =
-        mockMvcUtils.cloneFlexResource(
-            USER_REQUEST,
-            /*sourceWorkspaceId=*/ workspaceId,
-            sourceFlex.getResourceId(),
-            /*destWorkspaceId=*/ workspaceId2,
-            ApiCloningInstructionsEnum.RESOURCE,
-            destResourcename,
-            destDescription);
-
-    // Assert resource returned in clone flight response.
-    mockMvcUtils.assertClonedControlledFlexibleResource(
-        sourceFlex.getFlexibleResource(),
-        clonedFlexResource,
-        /*expectedDestWorkspaceId=*/ workspaceId2,
-        destResourcename,
-        destDescription,
-        USER_REQUEST.getEmail(),
-        USER_REQUEST.getEmail());
-
-    // Assert resource returned by get
-    final ApiFlexibleResource gotResource =
-        mockMvcUtils.getFlexibleResource(
-            USER_REQUEST, workspaceId2, clonedFlexResource.getMetadata().getResourceId());
-
-    assertApiFlexibleResourceEquals(clonedFlexResource, gotResource);
-  }
-
-  @Test
-  void clone_copyResource_undo() throws Exception {
-    UUID workspaceId = mockMvcUtils.createWorkspaceWithoutCloudContext(USER_REQUEST).getId();
-    UUID workspaceId2 = mockMvcUtils.createWorkspaceWithoutCloudContext(USER_REQUEST).getId();
-    var sourceFlex = createDefaultFlexResourceWithData(workspaceId, null);
-
-    String destResourceName = TestUtils.appendRandomNumber("dest-resource-name");
-    String destDescription = "new description";
-
-    mockMvcUtils.cloneFlex_undo(
-        USER_REQUEST,
-        /*sourceWorkspaceId=*/ workspaceId,
-        sourceFlex.getResourceId(),
-        /*destWorkspaceId=*/ workspaceId2,
-        ApiCloningInstructionsEnum.RESOURCE,
-        destResourceName,
-        destDescription);
-
-    // Assert clone doesn't exist. There's no resource ID, so search on resource name.
-    mockMvcUtils.assertNoResourceWithName(USER_REQUEST, workspaceId2, destResourceName);
-  }
 }
